Changed Organism to store a pointer to the CFG passed to its constructor. Removed CFG argument from Organism::setFitness.
Changed Organism constructor to no longer create a test suite. Instead I
added a initializeRandomChromosome method to do that. 

Idea is to have another method that allows the building of a chromosome
from other chromosomes via crossover.

Also added an initialized and evaluated flag to organism. Will be used
to assert the order in which these public methods are called. May or may
not be necessary if this can be done with sentinal values on the
chromosome and fitness properties.