renamed function, calculate coverage ratio was not as accurate as I thought.
Also, I added a call to updateCoverage in the replaceOrganism function.
For some reason the  updateCoverageBeforeReplacement was not working,
and the simulation refused to update its coverage ratio past the first
one.

I also renamed that replace organism function to a more general name
since it is not always replacing the child.