Implemented a finalTestSuite in the rangeSet.
Now usefulness counts are only updated when a generated test case covers
something new in this suite, at which point that new test case is added
to the suite before returning. 

I also added an offerToFinalTestSuite function which is called by the
tryLocalOpt function when its successful.

Thus at the end of the simulation, this test suite is a minimal suite
containing all the coverage of the entire population.

I think this solved the usefulness overflow issue which was a side
effect of not being able to update populationCoverage when we needed to
for a valid comparison.