Fixed two bugs in these control flow graphs. The code to sum the first and second halves was wrong. First there was a hardcoded value of 5 for half of the parameters, changed this to number of parameters. But then the array index for the sum of second half didn't make sense and always resulted in the halves being equal. I think you must have had it written differently then changed it but never debugged it and looked at the numbers.
In addition several blocks had the code 
	int a1 = testCase->getParameter(0),
		a2 = testCase->getParameter(0),
		a3 = testCase->getParameter(0);
Based on the context I assume you meant
	int a1 = testCase->getParameter(0),
		a2 = testCase->getParameter(1),
		a3 = testCase->getParameter(2);
So I changed it to that. We now get ~70% on on med ~96% on the other and
~78% on the hard. The majority of this addition coverage seems to be
thanks to the local opt from 0 again. Without it and just param local
opt we get up to about 48% on hard.

Havent looked far into it yet but I think theres a bunch of other unused
predicates too that are counted in total number of pred but never
referenced in the blocks. It looks like all the ones where its a single
T or F but the actual block has multiple condition coverage are never
used.