Found another typo that lead to unreachable edge.
Now the only edges we fail to cover in the hard and med CFGs are the
ones cooresponding to these checks

if ( sumOfFirstHalf == sumOfSecondHalf && sumOfAll > 0 ) {   

and

if( (a1 == a2) && (a2 == a3) ) {

Which makes sense because they are very hard to get. We actually do
cover some of them that are possible with the local opt around 0, but
the ones where they have to be large numbers and equal are just too
unlikely to get.