I believe I finally tracked down the error in the ranges thanks to valgrind. We were accessing usesArray[size] in the case where valueUsed == end. The assert statement there was ill written so it never triggered.
Another issue could have been we were violating the rule of 3 by not
implementing the assignment copy constructor for Range class.

Also found a memory leak in population, not deallocating the edges and
predicate counts.

Before I found the real issue I converted rangeSet to use a vector
instead of dynamic array. This probably wasn't needed but the code's a
hell of a lot cleaner now and hopefully won't cause further issues.

It looks like there is technically a memory leak in the control flow
graphs not deallocating the ranges for each parameter array. However we
only change CFGs a few times and this shouldn't be a huge deal. Still
might want to fix this.

There is one remaining issue that I can't figure out. Almost everywhere
the default test case constructor is called valgrind gives a definitely
lost memory error. I tried many things that don't seem to fix the issue.
I will post this on the crashes issue shortly.