Added avg number of seconds as a column in the final results table. We can then use these times to give the random searcher the same amount.
Also added code to count the number of test cases coming from each
possible source in the algorithm. This is now printed out as a comment
in the final results table. I figured it would be interesting to know
and we may want to include it in a table of some form but wasn't sure so
just made it a comment.

In addition I made a small change to the range code that should help
with the seemingly exponential increase in time it takes as the numbe
rof generations increases. I believe the issue was that the range set
always adds adjacent ranges towards the end when the std dev becomes so
low that many are above it. The logic used for adapting ranges is not
the best and basically leads to an explosion in the number of ranges the
longer the simulation goes on. A bunch of time is then spent sorting,
deleting, and readding basically the same ranges over and over again.
Now it atleast won't add any adjacent ranges or split the range unless
the existing range size / 2 > 25.